+++ lib/to_source.rb
@@ -1,6 +1,5 @@
 require 'set'
 require 'melbourne'
-require 'to_source/version'
 require 'to_source/visitor'
 
 # Namespace of library